# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RatingRequest": {
        "additionalProperties": false,
        "description": "Rate the other party on a completed order.",
        "properties": {
          "comment": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "score": {
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Score",
            "type": "integer"
          }
        },
        "required": [
          "score"
        ],
        "title": "RatingRequest",
        "type": "object"
      },
      "RatingResponse": {
        "description": "A created rating.",
        "properties": {
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "order_id": {
            "title": "Order Id",
            "type": "string"
          },
          "rated_user_id": {
            "title": "Rated User Id",
            "type": "string"
          },
          "score": {
            "title": "Score",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "order_id",
          "rated_user_id",
          "score"
        ],
        "title": "RatingResponse",
        "type": "object"
      },
      "RatingSummaryResponse": {
        "description": "A user's aggregate received rating.",
        "properties": {
          "average_score": {
            "title": "Average Score",
            "type": "string"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "average_score",
          "count"
        ],
        "title": "RatingSummaryResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "SAFE-GIFT API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/orders/{order_id}/ratings": {
      "post": {
        "description": "Rate the other party on a completed order.",
        "operationId": "rate_order_api_orders__order_id__ratings_post",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RatingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RatingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rate Order",
        "tags": [
          "ratings"
        ]
      }
    },
...
    "/api/users/{user_id}/ratings/summary": {
      "get": {
        "description": "Return a user's average received score and rating count.",
        "operationId": "user_rating_summary_api_users__user_id__ratings_summary_get",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RatingSummaryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "User Rating Summary",
        "tags": [
          "ratings"
        ]
      }
    },
...
  }
}